+++ src/rules/operation/RuleConditionalTransferLightOwnable2Step.sol
@@ -16,7 +16,11 @@ import {Ownable2StepERC165Module} from "../../modules/Ownable2StepERC165Module.s
  * @title RuleConditionalTransferLightOwnable2Step
  * @notice Ownable2Step variant of RuleConditionalTransferLight.
  */
-contract RuleConditionalTransferLightOwnable2Step is RuleConditionalTransferLightBase, Ownable2Step, Ownable2StepERC165Module {
+contract RuleConditionalTransferLightOwnable2Step is
+    RuleConditionalTransferLightBase,
+    Ownable2Step,
+    Ownable2StepERC165Module
+{
     /*//////////////////////////////////////////////////////////////
                              CONSTRUCTOR
     //////////////////////////////////////////////////////////////*/
@@ -42,8 +46,7 @@ contract RuleConditionalTransferLightOwnable2Step is RuleConditionalTransferLigh
         return Ownable2StepERC165Module.supportsInterface(interfaceId)
             || interfaceId == RuleEngineInterfaceId.RULE_ENGINE_INTERFACE_ID
             || interfaceId == ERC1404ExtendInterfaceId.ERC1404EXTEND_INTERFACE_ID
-            || interfaceId == RuleInterfaceId.IRULE_INTERFACE_ID
-            || interfaceId == type(IERC7551Compliance).interfaceId
+            || interfaceId == RuleInterfaceId.IRULE_INTERFACE_ID || interfaceId == type(IERC7551Compliance).interfaceId
             || interfaceId == type(IERC3643ComplianceFull).interfaceId;
     }