+++ src/rules/validation/abstract/core/RuleTransferValidation.sol
@@ -95,8 +95,7 @@ abstract contract RuleTransferValidation is
     function supportsInterface(bytes4 interfaceId) public view virtual returns (bool) {
         return interfaceId == RuleEngineInterfaceId.RULE_ENGINE_INTERFACE_ID
             || interfaceId == ERC1404ExtendInterfaceId.ERC1404EXTEND_INTERFACE_ID
-            || interfaceId == RuleInterfaceId.IRULE_INTERFACE_ID
-            || interfaceId == type(IERC7551Compliance).interfaceId
+            || interfaceId == RuleInterfaceId.IRULE_INTERFACE_ID || interfaceId == type(IERC7551Compliance).interfaceId
             || interfaceId == type(IERC3643IComplianceContract).interfaceId;
     }